+++ backend/src/controllers/UserController.ts
@@ -2,7 +2,7 @@ import { Request, Response } from "express";
 import { Model } from "sequelize";
 import UserServices from "../services/UserServices";
 import { MetricComparationModel} from "../services/MetricsService";
-import { createSession } from "../database/repositories/sessionRepository";
+import SessionRepository from "../database/repositories/SessionRepository";
 import MetricsService from "../services/MetricsService";
 
 interface IUserWithMetrics{
@@ -22,10 +22,15 @@ class UserController {
   async getAllUserInfo (req: Request, res: Response): Promise<Response<Model>> {
     try{
       const { username } = req.params;
-        
-      const user = await UserServices.getAllUserInfoByUserName(username);
+      const { userID , token, sessionType} = req.body;
+      if(await SessionRepository.checkSession(userID, token, sessionType)){
+
+        const user = await UserServices.getAllUserInfoByUserName(username);
             
-      return res.status(200).send(user); 
+        return res.status(200).send(user); }
+      else{
+        return res.status(489).send();
+      }
     } catch(error){
       if (error instanceof Error) return res.status(400).send({message: error.message});
             
@@ -43,15 +48,19 @@ class UserController {
   async getUserWithMetrics (req: Request, res: Response): Promise<Response<IUserWithMetrics>> {
     try{
       const { username } = req.params;
-        
-      const user = await UserServices.getUserAndMetrics(username);
-      const metricComparation = await MetricsService.metricsHistory(user.getDataValue("id"));
+      const { userID , token, sessionType} = req.body;
+      if(await SessionRepository.checkSession(userID, token, sessionType)){  
+        const user = await UserServices.getUserAndMetrics(username);
+        const metricComparation = await MetricsService.metricsHistory(user.getDataValue("id"));
 
-      const result = {} as IUserWithMetrics;
-      result.user = user;
-      result.metricsInfo = metricComparation;
+        const result = {} as IUserWithMetrics;
+        result.user = user;
+        result.metricsInfo = metricComparation;
       
-      return res.status(200).send(result);
+        return res.status(200).send(result);}
+      else{
+        return res.status(489).send();
+      }
     } catch(error){
       if (error instanceof Error) return res.status(400).send({message: error.message});
             
@@ -70,11 +79,11 @@ class UserController {
   async createUser (req: Request, res: Response): Promise<Response>{
     try{
       const { username } = req.params;
-      const { email, password }  = req.body;
+      const { email, password} = req.body;
       await UserServices.createUser(username, email, password);
 
       const userId = (await UserServices.getAllUserInfoByUserName(username)).get("id");
-      await createSession(userId as string);
+      await SessionRepository.createSession(userId as string);
 
       return res.status(200).send({message: "User created !"});
     }catch(error ){
@@ -94,11 +103,15 @@ class UserController {
   async updateUser (req: Request, res: Response): Promise<Response> {
     try {
       const { username } = req.params;
-      const { email, password }  = req.body;
+      const { userID , token, sessionType, email, password} = req.body;
+      if(await SessionRepository.checkSession(userID, token, sessionType)){
 
-      await UserServices.updateUser(username, email, password);
+        await UserServices.updateUser(username, email, password);
 
-      return res.status(200).send({message: "User updated !"});
+        return res.status(200).send({message: "User updated !"});}
+      else{
+        return res.status(489).send();
+      }
     } catch (error) {
       if (error instanceof Error) return res.status(400).send({message: error.message});
             
@@ -114,12 +127,33 @@ class UserController {
     */
   async deleteUser (req: Request, res: Response): Promise<Response> {
     try {
-      const { username } = req.body;
+      const { userID , token, sessionType, username } = req.body;
+      if(await SessionRepository.checkSession(userID, token, sessionType)){  
+        await UserServices.deleteUser(username);
 
-      await UserServices.deleteUser(username);
-
-      return res.status(200).send({message: "User deleted !"});
+        return res.status(200).send({message: "User deleted !"});}
+      else{
+        return res.status(489).send();
+      }
     } catch (error) {
+      if (error instanceof Error) return res.status(400).send({message: error.message});
+            
+      return res.status(400).send({message: error});
+    }
+  }
+
+  async resetPassword (req: Request, res: Response): Promise<Response> {
+    try{
+      const {email} = req.params;
+      const { userID , token, sessionType } = req.body;
+      if(await SessionRepository.checkSession(userID, token, sessionType)){  
+        await UserServices.resetPass(email);
+
+        return res.status(200).send({message: "User deleted !"});}
+      else{
+        return res.status(489).send();
+      }
+    }  catch (error) {
       if (error instanceof Error) return res.status(400).send({message: error.message});
             
       return res.status(400).send({message: error});
